Translation for invalid list ranges
